% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EssentialJSONSerializer.R
\docType{package}
\name{essentialjsonserializer}
\alias{essentialjsonserializer}
\alias{essentialjsonserializer-package}
\title{JSON serialization of complex R-objects \cr}
\description{
The intention of the package fuctionality is to (de)serialize language specific complex objects towards generic 
JSON-datatype objects. The JSON conversion allows for flexible cross language communication when the package is implemented
in different languages. This package provides an R implementation of the (de)serialization functionality.
}
\details{
Currently, 6 generic complex JSON-datatypes are provided. Corresponding R-mappings are defined as follows:

\itemize{
 \item JSON datetime         - maps to POSIXct objects
 \item JSON timeseries       - maps to 1-dimensional xts objects
 \item JSON timeseries_n_dim - maps to n-dimensional xts objects (n > 1)
 \item JSON series:          - maps to 1-dimensional dataframe objects
 \item JSON dataframe:       - maps to n-dimensional dataframe objects (n > 1)
 \item JSON matrix:          - maps to matrix objects
}

The package provides the user with the EssentialJSONSerializer reference class. Following functions
are provided:
\itemize{
 \item RToList: Convert complex R structures to a generic datatype list structure.
 \item RToJSON: Convert complex R structures to a generic datatype JSON structure.
 \item RFromList: Convert generic datatype list structure to complex R-specific objects.
 \item RFromJSON: Convert generic datatype JSON structure to complex R-specific objects.
}

Examples: \cr
 - https://github.com/VermeirJellen/essentialjsonserializer_r \cr
 - http://EssentialDataScience.com/blog/essentialjsonserializer \cr
}
